move some store properties into temporaryConfiguration (#1807)
