+++ app/assets/javascripts/oxalis/controller/combinations/minimal_skeletontracing_arbitrary_controller.js
@@ -8,7 +8,8 @@ import Store from "oxalis/store";
 import Constants from "oxalis/constants";
 import ArbitraryController from "oxalis/controller/viewmodes/arbitrary_controller";
 import { InputKeyboard, InputKeyboardNoLoop } from "libs/input";
-import { setFlightmodeRecordingAction, deleteNodeAction, createBranchPointAction, requestDeleteBranchPointAction } from "oxalis/model/actions/skeletontracing_actions";
+import { deleteNodeAction, createBranchPointAction, requestDeleteBranchPointAction } from "oxalis/model/actions/skeletontracing_actions";
+import { setFlightmodeRecordingAction } from "oxalis/model/actions/settings_actions";
 import { zoomInAction, zoomOutAction, yawFlycamAction, pitchFlycamAction } from "oxalis/model/actions/flycam_actions";
 import type Model from "oxalis/model";
 import type View from "oxalis/view";
