+++ app/assets/javascripts/oxalis/controller/viewmodes/arbitrary_controller.js
@@ -16,8 +16,8 @@ import type { ModeType, Vector3, Point2 } from "oxalis/constants";
 import Model from "oxalis/model";
 import View from "oxalis/view";
 import Store from "oxalis/store";
-import { updateUserSettingAction } from "oxalis/model/actions/settings_actions";
-import { setFlightmodeRecordingAction, setActiveNodeAction, deleteNodeAction, createTreeAction, createNodeAction, createBranchPointAction, requestDeleteBranchPointAction, setViewModeAction } from "oxalis/model/actions/skeletontracing_actions";
+import { updateUserSettingAction, setFlightmodeRecordingAction, setViewModeAction } from "oxalis/model/actions/settings_actions";
+import { setActiveNodeAction, deleteNodeAction, createTreeAction, createNodeAction, createBranchPointAction, requestDeleteBranchPointAction } from "oxalis/model/actions/skeletontracing_actions";
 import SceneController from "oxalis/controller/scene_controller";
 import SkeletonTracingController from "oxalis/controller/annotations/skeletontracing_controller";
 import { getBaseVoxel } from "oxalis/model/scaleinfo";
