+++ app/assets/javascripts/oxalis/model/actions/settings_actions.js
@@ -5,6 +5,7 @@
  * @flow
  */
 import type { UserConfigurationType, DatasetConfigurationType, DatasetLayerConfigurationType, TemporaryConfigurationType, DatasetType } from "oxalis/store";
+import type { ModeType } from "oxalis/constants";
 
 type UpdateUserSettingActionType = {type: "UPDATE_USER_SETTING", propertyName: $Keys<UserConfigurationType>, value: any};
 type UpdateDatasetSettingActionType = {type: "UPDATE_DATASET_SETTING", propertyName: $Keys<DatasetConfigurationType>, value: any};
@@ -13,6 +14,8 @@ type ToggleTemporarySettingActionType = {type: "TOGGLE_TEMPORARY_SETTING", prope
 type UpdateLayerSettingActionType = {type: "UPDATE_LAYER_SETTING", layerName:string, propertyName: $Keys<DatasetLayerConfigurationType>, value: any};
 export type InitializeSettingsAction = {type: "INITIALIZE_SETTINGS", initialUserSettings: UserConfigurationType, initialDatasetSettings: DatasetConfigurationType};
 type SetDatasetAction = {type: "SET_DATASET", dataset: DatasetType};
+type SetViewModeActionType = {type: "SET_VIEW_MODE", viewMode: ModeType};
+type SetFlightmodeRecordingActionType = {type: "SET_FLIGHTMODE_RECORDING", value: boolean};
 
 export type SettingActionType = (
   UpdateUserSettingActionType |
@@ -62,4 +65,14 @@ export const initializeSettingsAction = (initialUserSettings:Object, initialData
   type: "INITIALIZE_SETTINGS",
   initialUserSettings,
   initialDatasetSettings,
+});
+
+export const setViewModeAction = (viewMode: ModeType): SetViewModeActionType => ({
+  type: "SET_VIEW_MODE",
+  viewMode,
+});
+
+export const setFlightmodeRecordingAction = (value: boolean): SetFlightmodeRecordingActionType => ({
+  type: "SET_FLIGHTMODE_RECORDING",
+  value,
 });
