+++ app/assets/javascripts/oxalis/model/reducers/settings_reducer.js
@@ -66,6 +66,16 @@ function SettingsReducer(state: OxalisState, action: ActionType): OxalisState {
       });
     }
 
+    case "SET_VIEW_MODE": {
+      return update(state, {
+        temporaryConfiguration: { viewMode: { $set: action.viewMode } },
+      });
+    }
+    case "SET_FLIGHTMODE_RECORDING": {
+      return update(state, {
+        temporaryConfiguration: { flightmodeRecording: { $set: action.value } },
+      });
+    }
     default:
       // pass;
   }
