+++ app/assets/javascripts/oxalis/model/reducers/skeletontracing_reducer.js
@@ -59,17 +59,6 @@ function SkeletonTracingReducer(state: OxalisState, action: ActionType): OxalisS
       const newState = update(state, { tracing: { $set: skeletonTracing } });
       return zoomReducer(newState, contentData.zoomLevel);
     }
-    // TODO: these should probably live somewhere else
-    case "SET_VIEW_MODE": {
-      return update(state, {
-        viewMode: { $set: action.viewMode },
-      });
-    }
-    case "SET_FLIGHTMODE_RECORDING": {
-      return update(state, {
-        flightmodeRecording: { $set: action.value },
-      });
-    }
     default:
       // pass
   }
